requirements.txt
fastapi==0.104.1
pydantic==2.5.0
python-dotenv==1.0.0
uvicorn==0.24.0
requests==2.32.4
sentence-transformers==2.7.0
pdfplumber==0.10.3
google-generativeai==0.8.0
faiss-cpu==1.7.4
numpy==1.24.3
